organized as division and multiplication
